backend/requirements.txt
# SECaaS Insider Threat Detection - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0

# Validation
pydantic>=2.5.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

